requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
drf-yasg==1.21.7
django-environ==0.11.2
selenium==4.15.2
webdriver-manager==4.0.0 